# backend/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "企业知识库问答系统 API",
    "version": "1.0.0",
...
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "本地开发"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "UserId": {
        "type": "apiKey",
        "in": "header",
        "name": "x-user-id",
        "description": "登录用户的 MongoDB _id"
      },
      "UserName": {
        "type": "apiKey",
        "in": "header",
        "name": "x-user-name",
        "description": "登录用户名"
      },
      "UserRole": {
        "type": "apiKey",
        "in": "header",
        "name": "x-user-role",
        "description": "用户角色：admin 或 user"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "code": {
            "type": "integer"
          },
          "msg": {
            "type": "string"
          },
          "error": {
            "type": "string"
          }
        }
      },
...
      "LoginRequest": {
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "properties": {
          "username": {
            "type": "string",
            "example": "admin"
          },
          "password": {
            "type": "string",
            "example": "123456"
          }
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "登录成功"
          },
          "data": {
            "type": "object",
            "properties": {
              "_id": {
                "type": "string"
              },
              "username": {
                "type": "string"
              },
              "nickname": {
                "type": "string"
              },
              "role": {
                "type": "string"
              }
            }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/chat/admin/all": {
      "get": {
        "tags": [
          "Chat"
        ],
        "summary": "管理员查看所有会话",
        "security": [
          {
            "UserId": [],
            "UserName": [],
            "UserRole": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          },
          "403": {
            "description": "无权限"
          },
          "500": {
            "description": "服务器错误"
          }
        }
      }
    },
    "/api/chat": {
      "get": {
        "tags": [
          "Chat"
        ],
        "summary": "获取当前用户会话列表",
        "security": [
          {
            "UserId": [],
            "UserName": [],
            "UserRole": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "categoryId",
            "schema": {
              "type": "string"
            },
            "description": "按知识库分类筛选"
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          },
          "500": {
            "description": "服务器错误"
          }
        }
      }
    },
    "/api/chat/{id}": {
      "get": {
        "tags": [
          "Chat"
        ],
        "summary": "获取会话详情",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          },
          "404": {
            "description": "会话不存在"
          },
          "500": {
            "description": "服务器错误"
          }
        }
      },
      "delete": {
        "tags": [
          "Chat"
        ],
        "summary": "删除会话",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功"
          },
          "404": {
            "description": "会话不存在"
          },
          "500": {
            "description": "服务器错误"
          }
        }
      }
    },
    "/api/chat/save": {
      "post": {
        "tags": [
          "Chat"
        ],
        "summary": "手动保存会话",
        "description": "兼容手动存储会话，QA 流式完成后也可调用",
        "security": [
          {
            "UserId": [],
            "UserName": [],
            "UserRole": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "sessionId": {
                    "type": "string"
                  },
                  "categoryId": {
                    "type": "string"
                  },
                  "question": {
                    "type": "string"
                  },
                  "answer": {
                    "type": "string"
                  },
                  "sources": {
                    "type": "array",
                    "items": {
                      "type": "object"
                    }
                  },
                  "history": {
                    "type": "array",
                    "items": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "会话保存成功"
          },
          "500": {
            "description": "服务器错误"
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "用户登录",
        "description": "校验用户名密码，返回用户基础信息。前端登录后将信息存入 localStorage，后续请求通过 x-user-* 请求头携带身份。",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "登录成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "401": {
            "description": "用户名或密码错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "服务器错误"
          }
        }
      }
    },
...
  }
}